updated files

+++ part3/bufHash.C
@@ -13,7 +13,7 @@
 int BufHashTbl::hash(const File* file, const int pageNo)
 {
   int tmp, value;
-  tmp = (int)file;  // cast of pointer to the file object to an integer
+  tmp = (long)file;  // cast of pointer to the file object to an integer
   value = (tmp + pageNo) % HTSIZE;
   return value;
 }
